# docs/swagger.yaml
basePath: /api/v1
host: localhost:333
info:
  contact: {}
  description: Sensor API to maintain sensor data and generate aggregate statistics.
  title: Sensor API
  version: "1.0"
paths:
  /group/:groupName/species:
    get:
      description: Retrieves full list of species with counts currently detected inside
        the group.
      operationId: query-species
      parameters:
      - description: Group name
        in: path
        name: groupName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Retrieve full list of species inside the group
  /group/:groupName/species/top/:N:
    get:
      description: Retrieves list of top N species with counts currently detected
        inside the group.
      operationId: query-species-with-filter
      parameters:
      - description: Group name
        in: path
        name: groupName
        required: true
        type: string
      - description: Top N species count
        in: path
        name: "N"
        required: true
        type: integer
      - description: From time in Unix timestamp
        in: query
        name: from
        type: integer
      - description: Till time in Unix timestamp
        in: query
        name: till
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Retrieve list of top N species inside the group
  /group/:groupName/temperature:
    get:
      description: Collect average temperature
      operationId: query-average-temperature
      parameters:
      - description: Group name
        in: path
        name: groupName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Collect average temperature of sensors within a sensor group
  /group/:groupName/transparency:
    get:
      description: Collect average transparency
      operationId: query-average-transparency
      parameters:
      - description: Group name
        in: path
        name: groupName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Collect average transparency of sensors within a sensor group
  /region/temperature/max:
    get:
      description: Calculate maximum temperature inside a region
      operationId: calculate-max-temperature
      parameters:
      - description: Minimum x
        in: query
        name: xMin
        required: true
        type: number
      - description: Maximum x
        in: query
        name: xMax
        required: true
        type: number
      - description: Minimum y
        in: query
        name: yMin
        required: true
        type: number
      - description: Maximum y
        in: query
        name: yMax
        required: true
        type: number
      - description: Minimum z
        in: query
        name: zMin
        required: true
        type: number
      - description: Maximum z
        in: query
        name: zMax
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Calculate maximum temperature
  /region/temperature/min:
    get:
      description: Calculate minimum temperature inside a region
      operationId: calculate-min-temperature
      parameters:
      - description: Minimum x
        in: query
        name: xMin
        required: true
        type: number
      - description: Maximum x
        in: query
        name: xMax
        required: true
        type: number
      - description: Minimum y
        in: query
        name: yMin
        required: true
        type: number
      - description: Maximum y
        in: query
        name: yMax
        required: true
        type: number
      - description: Minimum z
        in: query
        name: zMin
        required: true
        type: number
      - description: Maximum z
        in: query
        name: zMax
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Calculate minimum temperature
  /sensor/:codeName/temperature/average:
    get:
      description: Calculate average temperature in a given time interval by a sensor
      operationId: calculate-average-temperature-by-sensor
      parameters:
      - description: From time in Unix timestamp
        in: query
        name: from
        required: true
        type: integer
      - description: Till time in Unix timestamp
        in: query
        name: till
        required: true
        type: integer
      - description: Code name of the sensor
        in: path
        name: codeName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Calculates average temperature
swagger: "2.0"